修正：モーダルの画面サイズ調整、モーダル関係の変数名の変更
