+++ frontend/src/types/happiness-all-response.ts
@@ -24,7 +24,6 @@ export interface MapDataItem {
   }
   answers: HappinessFields
   memos: string[]
-  totalmemos: string
 }
 
 type GraphData = {
